pyproject.toml
[tool.poetry]
name = "fastapi-starter"
version = "0.1.0"
description = ""
authors = ["Timo Wilhelm <mail@timowilhelm.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
pydantic = { extras = ["dotenv"], version = "^1.10.2" }
fastapi = "^0.85.1"
uvicorn = { extras = ["standard"], version = "^0.19.0" }
gunicorn = "^20.1.0"
slowapi = "^0.1.6"
httpx = "^0.23.0"
redis = "^4.3.4"
cryptography = "^38.0.3"
pyjwt = "^2.6.0"
asyncpg = "^0.27.0"
sqlalchemy = "2.0.0b3"
alembic = "^1.8.1"
opentelemetry-instrumentation-logging = "^0.35b0"
opentelemetry-instrumentation-fastapi = "^0.35b0"
opentelemetry-instrumentation-httpx = "^0.35b0"
opentelemetry-instrumentation-redis = "^0.35b0"
opentelemetry-instrumentation-sqlalchemy = "^0.35b0"
opentelemetry-instrumentation-asyncpg = "^0.35b0"
opentelemetry-instrumentation-system-metrics = "^0.35b0"
azure-monitor-opentelemetry-exporter = "^1.0.0b8"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
flake8-bugbear = "^22.10.27"
pre-commit = "^2.20.0"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
color = false             # color output breaks vscode python extension
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.pyright]
reportGeneralTypeIssues = false

[tool.coverage.run]
source = ["app"]
omit = ["app/gunicorn_worker.py"]
